Update build-configuration.md
This change demostrates how you can include a multi-line PowerShell script *inside* an event (in order that other Powershell / CMD lines could be used before / after it).  I noticed that the extra line breaks appeared to be required in my own usage and so have documented it as such.  

For what it's worth I've had problems using the multi-line PowerShell script covering an entire event that is already documented.  As a result I simply don't use that method now that I know there is this alternate approach.