Clarify order and behavior related to directory changes (#504)
It seems logical that the test steps continue in the same directory
where the build left off, but this was not explicit. Even more
confusing, "build.project" is *not* affected by directory changes in
"before_build".

I am not sure when "before_package" is exactly executed, so it remains
undocumented in build-configuration.md.