Update services-databases.md (#527)
* Update services-databases.md

I propose a replacement for the script to set SQL instances to use dynamic ports. This script enumerates the registry to find currently installed instances and sets them to dynamic ports. I did this as I noted that the workers I'm getting have only 2014, 2016 and 2017 - which will eventually change, so a dynamic solution future-proofs it.
Example here: https://ci.appveyor.com/project/fireflycons/invoke-sqlexecute

* Update services-databases.md

Update as requested by Feodor.

Question - Are the currently available SQL servers now only 2014, 2016 and 2017, or do different images still have earlier versions? If not the page should also be updated to reflect that.

Thanks

* Fix build erroros