Allow connection name to be something other than DefaultConnection (#781)
The documentation said that the parameter and environment variable name has to be `DefaultConnection-Web.config Connection String`. This isn't necessarily the case. It has to be the name of your database connection + `-Web.config Connection String`.

So if you have this in your Web.config:

```
<connectionStrings>
    <add name="MainDb" connectionString="..." providerName="System.Data.SqlClient" />
</connectionStrings>
```

Then you can use `MainDb-Web.config Connection String` as parameter name and environment variable name.